use year over phase number in title
I think date(s) in title is better because it is an absolute value. When
I see proposal title like "Decred Submarine Phase 9", the "9" is not as
informative as "2027". Seeing year or date immediately tells me the time
context.

But this is my opinion, I'm ok if proposal author has own style and uses
phases over dates.