Phase 2 fixes 2
* fix digit character

Change English capital O to a zero 0 character.

Monospace font helps detect these :)

* fix thousand separator

Different countries use different thousand separator like 1,000 or
1.000. Change to comma for consistency with most Politeia proposals
and the rest of this doc.

* improve table formatting

Align all numbers to the right, they are easier to read this way. Also
helps quickly detect different orders of magnitude when the width
differs:

      $100
    $1,000

* add total row

I think it makes the table even faster to understand. Remove if not.

* clarify budget line

* use sentence case in table details

* use year over phase number in title

I think date(s) in title is better because it is an absolute value. When
I see proposal title like "Decred Submarine Phase 9", the "9" is not as
informative as "2027". Seeing year or date immediately tells me the time
context.

But this is my opinion, I'm ok if proposal author has own style and uses
phases over dates.

* drop Creation from title

Shorter title is faster reading. "Creation" is a beautiful word though.

Feel free to revert if you don't like this change!

* use last day of Nov 2022 in end date

The "End Date" has no strict meaning yet. Currently I read it as "last
day of billable work" so let's make it 1 day less. Including Dec 1 in
the date range means you could technically do billable work on that day
and submit an invoice for one day.

This is just my thing, people who approve invoices may not read it like
this.

* mention that upcoming event is not in the report yet

* add item numbers to Work included in this proposal

It had 3 paragraphs but 2 logical items. Transition to 3rd paragraph
felt odd. I added numbers to group the paragraphs.

* simplify phase 1 report intro

I could not find a wording for end date/remaining budget that would not
feel awkward, so just remove it and defer to first proposal's text
for detailed conditions.