Proposal phase 2
* Create proposal-phase-2.md

* Update proposal-phase-2.md

* Phase 2 fixes 1 (#6)

* fix link syntax

* fix table formatting

* swap Spent and Allocated columns

* add Total column

* correct structure

Change headings to make the following high-level structure:

- Summary
- Phase 1 report
  - Events
  - Media presence and content
  - Social media
  - Summary and finances
- Work included in this proposal
- Who?

* simplify links

For archived Politeia proposals we can use proposals.decred.org and it
will redirect to proposals-archive.decred.org. When old proposals are
migrated to the new site, this link will not break.

For decredcommunity events we can link to published versions of the
reports.

* add spacing between list items for readability

* move events para under Events

* text tweaks

* add SM links

* round SM metrics and add diffs

* move first prop link under phase 1 report

* reword phase 1 report intro and finances para

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Phase 2 fixes 2

* fix digit character

Change English capital O to a zero 0 character.

Monospace font helps detect these :)

* fix thousand separator

Different countries use different thousand separator like 1,000 or
1.000. Change to comma for consistency with most Politeia proposals
and the rest of this doc.

* improve table formatting

Align all numbers to the right, they are easier to read this way. Also
helps quickly detect different orders of magnitude when the width
differs:

      $100
    $1,000

* add total row

I think it makes the table even faster to understand. Remove if not.

* clarify budget line

* use sentence case in table details

* use year over phase number in title

I think date(s) in title is better because it is an absolute value. When
I see proposal title like "Decred Submarine Phase 9", the "9" is not as
informative as "2027". Seeing year or date immediately tells me the time
context.

But this is my opinion, I'm ok if proposal author has own style and uses
phases over dates.

* drop Creation from title

Shorter title is faster reading. "Creation" is a beautiful word though.

Feel free to revert if you don't like this change!

* use last day of Nov 2022 in end date

The "End Date" has no strict meaning yet. Currently I read it as "last
day of billable work" so let's make it 1 day less. Including Dec 1 in
the date range means you could technically do billable work on that day
and submit an invoice for one day.

This is just my thing, people who approve invoices may not read it like
this.

* mention that upcoming event is not in the report yet

* add item numbers to Work included in this proposal

It had 3 paragraphs but 2 logical items. Transition to 3rd paragraph
felt odd. I added numbers to group the paragraphs.

* simplify phase 1 report intro

I could not find a wording for end date/remaining budget that would not
feel awkward, so just remove it and defer to first proposal's text
for detailed conditions.

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Phase 2 fixes 3

* nits

* simplify budget para + explicit monthly

* move closing words down

Expense breakdown looks better immediately after its summary paragraph.

Put "feedback" sentence after acknowledgements sentence. This way call
for feedback is closer to the end and has more chance to stay in
reader's attention.

* add youtube stat delta

* rework intro

Mention of prior work in 2018-2020 gives more context that you've been
around for a long time.

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

* Update proposal-phase-2.md

Co-authored-by: xaur <24484727+xaur@users.noreply.github.com>